% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTSSR2.R
\name{scTSSR2}
\alias{scTSSR2}
\title{use scTSSR2 to impute dropout values in scRNA-seq data}
\usage{
scTSSR2(X.count, k.gene = NULL, k.cell = NULL, W = NULL,
  lambda = 256, percent = 0.05, ncores = 1, MAX.ITER = 4,
  ABSTOL = 0.001, learning.rate = 1e-04, epochs = 100,
  verbose = TRUE, estimates.only = FALSE)
}
\arguments{
\item{X.count}{Raw read count matrix. The rows correspond to genes and the columns correspond to cells.
Can be sparse.}

\item{k.gene}{A hyper-parameter that controls the sparsity level of the estimated coefficient matrices, A1 and A2.
Default is k_{gene} = min(100, m/30).}

\item{k.cell}{A hyper-parameter that controls the sparsity level of the estimated coefficient matrices, B1 and B2.
Default is k_{cell} = min(100, n/30).}

\item{W}{A weight matrix with element W_{gc} denotes the non-dropout probability of the expression level of gene g in cell c.
Default is W_{gc}=X_{gc}/max(X_{gc}).}

\item{lambda}{Ridge penalty parameter. Default is 256.}

\item{percent}{The expression count matrix is preprocessed by filtering out the genes
expressed in at most percent*\eqn{100\%} of the cells. Default is 0.05.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{MAX.ITER}{Maximum iteration of the external circulation of scTSSR2. Default is 4.}

\item{ABSTOL}{Absolute tolerance of the external circulation. Default is 1e-3.}

\item{learning.rate}{A hyper-parameter that controls the speed of adjusting the weights of the network with respect to the loss gradient.
Default is 0.0001.}

\item{epochs}{The number of the entire training set going through the entire network. Default is 100.}

\item{verbose}{Whether to output the value of metrics at the end of each epoch. Default is TRUE.}

\item{estimates.only}{logical item. A logical flag to determine whether to output only imputed estimates.
Default is FALSE.}
}
\value{
If `estimates.only = TRUE', then a matrix of scTSSR2 estimates.

If `estimates.only = FALSE', a list with the following components

\item{\code{estimate}}{Recovered (normalized) expression.}

\item{\code{se}}{Standard error of estimates.}

\item{\code{info}}{Information about dataset.}


The \code{info} element is a list with the following components:

\item{\code{size.factor}}{Size factor used for normalization.}

\item{\code{pred.time}}{Time taken to generate predictions.}

\item{\code{posterior.time}}{Time taken to compute the posterior distribution.}

\item{\code{total.time}}{Total time for scTSSR2 estimation.}
}
\description{
use scTSSR2 to impute dropout values in scRNA-seq data
}
\examples{

data("baron")

baron_imputation_result = scTSSR2(baron$count.samp)

}
\author{
Ke Jin, \email{kej13@mails.ccnu.edu.cn}
}
